requirements_pip.txt
torch-scatter==2.0.4
torch-sparse==0.6.1
torch-cluster==1.5.4
torch-spline-conv==1.2.0
descartes==1.1.0
fire==0.3.1
jupyter==1.0.0
opencv_python==4.3.0.*
Shapely==1.7.0
Pillow==6.2.1
torch_geometric==1.6.0
open3d
